+++ crates/hyperqueue/src/transfer/messages.rs
@@ -15,6 +15,7 @@ use std::path::PathBuf;
 use std::time::Duration;
 use tako::gateway::{LostWorkerReason, ResourceRequestVariants, TaskDataFlags, WorkerRuntimeInfo};
 use tako::program::ProgramDefinition;
+use tako::server::TaskExplanation;
 use tako::worker::WorkerConfiguration;
 use tako::{JobId, JobTaskCount, JobTaskId, Map, TaskId, WorkerId};
 
@@ -36,6 +37,7 @@ pub enum FromClientMessage {
     ServerInfo,
     OpenJob(JobDescription),
     CloseJob(CloseJobRequest),
+    TaskExplain(TaskExplainRequest),
 
     // This command switches the connection into streaming connection,
     // it will no longer reacts to any other client messages
@@ -252,6 +254,13 @@ pub struct JobInfoRequest {
 }
 
 #[derive(Serialize, Deserialize, Debug)]
+pub struct TaskExplainRequest {
+    pub job_selector: SingleIdSelector,
+    pub task_id: JobTaskId,
+    pub worker_id: WorkerId,
+}
+
+#[derive(Serialize, Deserialize, Debug)]
 pub struct JobDetailRequest {
     pub job_id_selector: IdSelector,
     pub task_selector: Option<TaskSelector>,
@@ -323,6 +332,13 @@ pub struct JobDetailResponse {
     pub server_uid: String,
 }
 
+#[derive(Serialize, Deserialize, Debug)]
+pub struct TaskExplainResponse {
+    pub task_id: TaskId,
+    pub worker_id: WorkerId,
+    pub explanation: TaskExplanation,
+}
+
 #[derive(Serialize, Deserialize, Debug, Clone)]
 pub struct ServerInfo {
     pub server_uid: String,
@@ -354,6 +370,7 @@ pub enum ToClientMessage {
     CloseJobResponse(Vec<(JobId, CloseJobResponse)>),
     Error(String),
     ServerInfo(ServerInfo),
+    TaskExplain(TaskExplainResponse),
     Event(Event),
     // This indicates in live event streaming when old events where
     // old streamed, and now we are getting new ones