+++ crates/tako/src/internal/server/mod.rs
@@ -3,6 +3,7 @@ pub mod comm;
 pub mod core;
 pub(crate) mod dataobj;
 pub(crate) mod dataobjmap;
+pub(crate) mod explain;
 pub mod reactor;
 pub mod rpc;
 pub mod task;