+++ crates/tako/src/internal/server/worker.rs
@@ -265,11 +265,11 @@ impl Worker {
     pub fn new(
         id: WorkerId,
         configuration: WorkerConfiguration,
-        resource_map: ResourceMap,
+        resource_map: &ResourceMap,
+        now: Instant,
     ) -> Self {
-        let resources = WorkerResources::from_description(&configuration.resources, &resource_map);
+        let resources = WorkerResources::from_description(&configuration.resources, resource_map);
         let load = WorkerLoad::new(&resources);
-        let now = std::time::Instant::now();
         Self {
             id,
             termination_time: configuration.time_limit.map(|duration| now + duration),